+++ src/controller/NivelAcessoController.java
@@ -5,97 +5,20 @@
  */
 package controller;
 
-import fabricas.AbstractFactory;
-import interfaces.Tabela;
+import dao.NivelAcessoDao;
 import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import javax.swing.JOptionPane;
 import model.NivelAcesso;
-import util.CriaStatement;
 
 /**
  *
  * @author fdz
  */
 public class NivelAcessoController extends Controller<NivelAcesso> {
-    
+
     public NivelAcessoController(Connection con) {
-
+        
         super(con);
-        
-        Tabela obj = AbstractFactory.getInstance("HUMANO").getTabela("NIVEL_ACESSO");
-        
-        this.id = obj.getNomeId();
-        this.tabela = obj.getNomeTabela();
-        this.criaStatement = new CriaStatement(con, tabela, id);
-        campos = "nac_descricao";
-        vetorCampos = campos.split(",");
-
-    }
-
-    @Override
-    public PreparedStatement statementInserir(NivelAcesso item) {
-        
-        try {
-
-            ps = criaStatement.insertSql(tabela, campos);
-
-            ps.setString(1, item.getDescricao());
-
-        } catch (Exception error) {
-
-            JOptionPane.showMessageDialog(null, error.getMessage());
-
-        }
-
-        return ps;
+        dao = new NivelAcessoDao(con);
         
     }
-
-    @Override
-    public PreparedStatement statementAlterar(NivelAcesso item) {
-        
-        try {
-
-            ps = criaStatement.updateSql(campos);
-
-            ps.setString(1, item.getDescricao());
-            ps.setInt(2, item.getId());
-
-        } catch (Exception error) {
-
-            JOptionPane.showMessageDialog(null, error.getMessage());
-
-        }
-
-        return ps;
-        
-    }
-
-    @Override
-    public NivelAcesso criaItem(ResultSet rs) {
-        
-        try {
-            
-            if (rs.next()) {
-
-                return new NivelAcesso(
-                        rs.getInt(id), 
-                        rs.getString(vetorCampos[0])
-                );
-
-            } else
-                throw new Exception("NivelAcesso não encontrado");
-            
-        } catch (Exception e) {
-            
-            JOptionPane.showMessageDialog(null, e.getMessage());
-            
-        }
-        
-        return null;
-        
-    }
-
 }
